add MPI mode to mercury-runner.   this allows both sides of a mercury-runner test to be launched with a single "mpirun" command (rather than having to run two commands).
enable this with cmake -DMPI=ON at config time.

details:

for MPI mode: we must be run in an MPI world with exactly 2 procs.
(in this case MPI is being used to launch mercury-runner but may
not be used for transport..)    rank 0 becomes the "local" proc
and rank 1 becomes the "remote" proc.  we assume that mpirun
has been told to use the hosts that match the local and remote
specs on the command line.

bidirectional:
  mpirun -n 2 ./mercury-runner -c 3 -l 1 -M -m cs \
             1 bmi+tcp://localhost:5555 bmi+tcp://localhost:5556

single direction:
  mpirun -n 2 ./mercury-runner -c 3 -l 1 -M -m c \
             1 bmi+tcp://localhost:5555 bmi+tcp://localhost:5556

in both cases rank 1 swaps the local and remote specs.

for single direction, rank 1 has its mode toggled from the given -m
value.  also, you must specify both remote and local specs when
using MPI mode (even for "-m s").